support floats as parameter for --wait
git-svn-id: http://arthurdejong.org/svn/webcheck/webcheck@259 86f53f14-5ff3-0310-afe5-9b438ce3f40c

+++ webcheck.py
@@ -111,7 +111,7 @@ def parse_args(site):
             elif flag in ('-r', '--redirects'):
                 config.REDIRECT_DEPTH = int(arg)
             elif flag in ('-w', '--wait'):
-                config.WAIT_BETWEEN_REQUESTS = int(arg)
+                config.WAIT_BETWEEN_REQUESTS = float(arg)
             elif flag in ('-V', '--version'):
                 print_version()
                 sys.exit(0)
